Fixed some MSVC code analysis warnings

+++ Graphics/GraphicsEngineD3D11/include/ShaderResourceCacheD3D11.hpp
@@ -458,11 +458,11 @@ private:
 
     static constexpr size_t MaxAlignment = std::max(std::max(std::max(alignof(CachedCB), alignof(CachedResource)), alignof(CachedSampler)), alignof(IUnknown*));
 
-    static constexpr Uint32 FirstCBOffsetIdx  = 0;                                      // | VS CB  | PS CB  | GS CB  | HS CB  | DS CB  | CS CB  |
-    static constexpr Uint32 FirstSRVOffsetIdx = FirstCBOffsetIdx + NumShaderTypes;      // | VS SRV | PS SRV | GS SRV | HS SRV | DS SRV | CS SRV |
-    static constexpr Uint32 FirstSamOffsetIdx = FirstSRVOffsetIdx + NumShaderTypes;     // | VS Sam | PS Sam | GS Sam | HS Sam | DS Sam | CS Sam |
-    static constexpr Uint32 FirstUAVOffsetIdx = FirstSamOffsetIdx + NumShaderTypes;     // | VS UAV | PS UAV | GS UAV | HS UAV | DS UAV | CS UAV |
-    static constexpr Uint32 MaxOffsets        = FirstUAVOffsetIdx + NumShaderTypes + 1; // | Count  |
+    static constexpr size_t FirstCBOffsetIdx  = 0;                                      // | VS CB  | PS CB  | GS CB  | HS CB  | DS CB  | CS CB  |
+    static constexpr size_t FirstSRVOffsetIdx = FirstCBOffsetIdx + NumShaderTypes;      // | VS SRV | PS SRV | GS SRV | HS SRV | DS SRV | CS SRV |
+    static constexpr size_t FirstSamOffsetIdx = FirstSRVOffsetIdx + NumShaderTypes;     // | VS Sam | PS Sam | GS Sam | HS Sam | DS Sam | CS Sam |
+    static constexpr size_t FirstUAVOffsetIdx = FirstSamOffsetIdx + NumShaderTypes;     // | VS UAV | PS UAV | GS UAV | HS UAV | DS UAV | CS UAV |
+    static constexpr size_t MaxOffsets        = FirstUAVOffsetIdx + NumShaderTypes + 1; // | Count  |
 
     std::array<OffsetType, MaxOffsets> m_Offsets = {};
